refactor: extract engine_capabilities module
